Move to unified graph model
Previously the registry was a special concept. We now have one compound
resource, the ResourceGraph, which is both a valid resource and a
registry. You can Register(string, Resource) and
RegisterDependency(from, to Resource).

While it would be completely valid to treat the composition of
ResourceGraphs as identical to the composition of leaf resources, we
perform a flatten operation when we find ourselves composing a
ResourceGraph onto another ResourceGraph. This results in us never
having a nested ResourceGraph (unless other, third party, composite
resources are introduced). This will be useful when it comes to
visualisation and suchlike.