Change Initialize to SetName
When dealing with nested composite resources (i.e. resource graphs
containing resource graphs), it's not useful to just see the name of the
leaf resource that is being executed in loggers and suchlike. A solution
to this is to provide each resource with the full path. When we register
a resource graph, we will update all of the child resources' names to
prepend the parent's name. As naming the resource graph happens after
naming the children, this requires calling Initialize multiple times,
which makes no sense. So, this diff also renames Intialize to SetName,
to avoid that issue.