Remove reinitialization of ResourceGraph derived fields
Previously SetName (then called Initialize) was called only once. As
such, it made sense to initialize fields such as dependencies with a
simple `rg.dependencies = map[Resource][]Resource{}`. However, this
single call invariant was removed when Initialize was renamed to
SetName. At this point I forgot to update the init function to handle
multiple calls, and broke ResourceGraph flattening in doing so.

This commit fixes that issue and adds a test case that aims to catch
issues with ResourceGraph flattening.