Modify Resource intefrace to Materialize(context.Context, chan<- Signal)
This is probably a mistake, but I want to preserve it so I can look at
it whenever I'm feeling cocky and be brought down to reality.

The basic problem I was trying to solve is optional resource runs. When
I have a config file, I probably want to reload the service that reads
that config file when it is modified. However, if the config file isn't
modified, we don't want to reload the service. This is what a normal
relation does:

```
rg := rfsb.ResourceGraph{}
configFile := &rfsb.FileResource{...}
rg.Register("configFile", configFile)
reload := &systemd.DaemonReload{...}
rg.When(configFile).Do(reload)
```

This causes the reload resource to be materialized always.

To solve this, I introduced the notion of Signals. Instead of always
running `reload` when `configFile`, we can run `reload` when
`configFile` emits a SignalMaterialized (as opposed to a SignalSkipped)
event. This makes our example look like:

```
rg := rfsb.ResourceGraph{}
configFile := &rfsb.FileResource{...}
rg.Register("configFile", configFile)
reload := &systemd.DaemonReload{...}
rg.When(configFile, SignalMaterialized).Do(reload)
```

However, this introduces other issues. For one, the Resource API gets
more complex:

```
type Resource interface {
	Materialize(context.Context, chan<- Signal) error

	// Best to implement these by embedding ResourceMeta
	Name() string
	SetName(string)
	Logger() *logrus.Entry
}
```
This was already an ugly interface, and now we're passing a random
channel to the Materialize function.

I also added implicit emissions to the ResourceGraph implementation.
This means that there are two kinds of signals; implicit and explicit.
If you are a user that has only used the built in resources (all of
which do not explicitly emit any singlas), you might think you needed to
explicitly emit the implicit signals, making it harder to work with.

On top of that, there is an issue with the way the resources chain.
Imagine the following graph:

```
rg := rfsb.ResourceGraph{}
configFile := &rfsb.FileResource{...}
rg.Register("configFile", configFile)
reload := &systemd.DaemonReload{...}
rg.When(configFile, SignalMaterialized).Do(reload)
restart := &systemd.RestartUnit{...}
rg.When(reload).Do(restart)
```

The `restart` resource is now implicitly an 'optional' resource. How
should we handle this? If the reload resource isn't run, we don't want
to wait for the second to run in order to run any others. My current
thought for this is that we should have a few built in singnals:

 - Evaluated - The resource's dependencies have been met
 - Skipped - The resource's dependencies have been met, the
      ShouldSkip function called and the result is true
 - Materialized - The resource's dependencies have been met and the
      Materialize function has been run
 - Finished - The resource's dependencies have finished, and the
      resource has been skipped/materialized

All these signals are emitted _after_ any calls to ShouldSkip or
Materialize. In this situation, we would change the above to

```
rg := rfsb.ResourceGraph{}
configFile := &rfsb.FileResource{...}
rg.Register("configFile", configFile)
reload := &systemd.DaemonReload{...}
rg.When(configFile, SignalMaterialized).Do(reload)
restart := &systemd.RestartUnit{...}
rg.When(reload, SignalEvaluated).Do(restart)
```

Notice that we use SignalEvaluated, not SignalFinished. This means that
if configFile does not Materialize, we will not Materialize either
reload or retart.

However, this is not right now. This commit is a flawed version (with
deadlocks). Enjoy!