Make signals a ResourceGroup level abstraction
In the previous commit, the Materialize function had a chan<- Signal
argument added, and the ResourceGroup's When function gained a Signal
argument. This commit reverts the change to Materialize, but keeps the
When component. Why? Simplicity.

Instead of having arbitrary signals, there are now 5 predefined signals
defined in signal.go. These are automatically emitted inside
ResourceGraph, based on the behavior of the resource (or the behaviour
of it's dependencies). The user can configure dependencies between
Resources using the signals via the When api.

There's a clear need for non binary dependencies between Resources,
which this diff solves. In the future, users may want more complex
signals, and some will potentially be added (NotMaterialized strikes me
as something that may be requested at some point). Even more complex
signals can still be done, but they will likely require the user to get
inventive. But at the end of the day, the user gets to write Go
functions. There's quite a lot of power right there.

This diff is much much simpler than the previous diff. If you don't like
signals, you can roll your own resource graph. That wasn't the case with
the previous diff.